More small changes
Removed the border around the outside of the content boxes.

Added Classes page to Navbar - think this could be good to implement tables.

Added some actual content to replace the placeholder text on the homepage.

Removed defunct Javascript link from bottom of index.html (was for testing a framework).

Moved background to body, seems to look better imo.