Major Content Changes
Added multiple pages for the content of the site. The more I was doing this I got slightly more skeptical of the possibility of marks being removed because the ISE already has a website, and it was in the practical for weeks 3 and 4.

I added some useful links to the side with Font-Awesome bullets just to make them a bit less bland.